Movement classification
Given the range of     orientation_x  orientation_y  orientation_z
amin    -179.949776     -81.820960    -179.839195
amax     179.873469      89.697391     179.981714 for Eulean orientation, where values near to -1 or 1 would indicate vertical movement, we set -90/90 as vertical movement thresholds that mirrows -1/1 in our dataset. If accelerometer data is not equal 0 we append "moving" and refine upon having glyroscope not equal 0 to turning/tilting. If glyroscope is indeed equal 0 for xyz, indicating no rotation, the thresholds for orientation can serve as classifiers for vertical vs. horozontal movement. If all conditions are not met, stationary is appended.