BERT
I am trying to build an inference pipeline with Google's Bert as Omotola suggested. This natural language inference will just output most probable words filling a gap in a sentence. But they can be used to train a neural network based on initial predictions. Later, the users' definitions will fine-tune both the neural network and BERT. For that, the cloud is required. For the inference, at least for API usage, a subscription is required.