Update Global_model.ipynb
Here, a global model with all user data could be trained, saved, and uploaded to a cloud service for easy deployment. If no user data is available yet, clustering will be skipped. The elbow point calculation ensures that we use the optimal number of clusters and not an arbitrary chosen pre-defined cluster number.